backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
exifread>=3.0.0
gunicorn>=21.0.0
python-decouple>=3.8
psycopg2-binary>=2.9.0
pytest>=7.0.0
pytest-django>=4.5.0
dj-database-url>=2.0.0